:truck: Move `create_function_from_source` from `nipype` to `CPAC`
